Building out asset publishing
