--- OpenGarage/html/publish_assets.sh
@@ -1,6 +1,0 @@
-./publish_assets ap_home.html html_ap_home
-./publish_assets sta_home.html html_sta_home
-./publish_assets sta_options.html html_sta_options
-./publish_assets sta_update.html html_sta_update
-./publish_assets sta_logs.html html_sta_logs
-./publish_assets porta.html portal
