+++ OpenGarage/www_assets.h
@@ -1,0 +1,1 @@
+const char WWW_ASSETS["html_ap_home"] PROGMEM = R"(<body><caption><b>OpenGarage WiFi Config</caption><br><br><table cellspacing=4 id='rd'>onfig</caption><br><br><tr><td>(Scanning...)</td></tr>fig</caption><br><br></table>(Scanning...)</td></tr>fig</caption><br><br><table cellspacing=16>/td></tr>fig</caption><br><br><tr><td><input type='text' name='ssid' id='ssid'></td><td>(SSID)</td></tr><tr><td><input type='password' name='pass' id='pass'></td><td>(Password)</td></tr><tr><td><input type='text' name='auth' id='auth'></td><td>(Auth Token)</td></tr>r><tr><td colspan=2><p id='msg'></p></td></tr>uth'></td><td>(Auth Token)</td></tr>r><tr><td><button type='button' id='butt' onclick='sf();' style='height:36px;width:180px'>Submit</button></td><td></td></tr></table><button type='button' id='butt' onclick='sf();' style='height:36px;width:180px'>Submit</button></td><td></td></tr><script><button type='button' id='butt' onclick='sf();' style='height:36px;width:180px'>Submit</button></td><td></td></tr>function id(s) {return document.getElementById(s);}();' style='height:36px;width:180px'>Submit</button></td><td></td></tr>function sel(i) {id('ssid').value=id('rd'+i).value;});' style='height:36px;width:180px'>Submit</button></td><td></td></tr>var tci; sel(i) {id('ssid').value=id('rd'+i).value;});' style='height:36px;width:180px'>Submit</button></td><td></td></tr>function tryConnect() {id').value=id('rd'+i).value;});' style='height:36px;width:180px'>Submit</button></td><td></td></tr>var xhr=new XMLHttpRequest();alue=id('rd'+i).value;});' style='height:36px;width:180px'>Submit</button></td><td></td></tr>xhr.onreadystatechange=function() {d('rd'+i).value;});' style='height:36px;width:180px'>Submit</button></td><td></td></tr>if(xhr.readyState==4 && xhr.status==200) {i).value;});' style='height:36px;width:180px'>Submit</button></td><td></td></tr>var jd=JSON.parse(xhr.responseText);200) {i).value;});' style='height:36px;width:180px'>Submit</button></td><td></td></tr>if(jd.ip==0) return;r.responseText);200) {i).value;});' style='height:36px;width:180px'>Submit</button></td><td></td></tr>var ip=''+(jd.ip%256)+'.'+((jd.ip/256>>0)%256)+'.'+(((jd.ip/256>>0)/256>>0)%256)+'.'+(((jd.ip/256>>0)/256>>0)/256>>0);/tr>id('msg').innerHTML='<b><font color=green>Connected! Device IP: '+ip+'</font></b><br>Device is rebooting. Switch back to<br>the above WiFi network, and then<br>click the button below to redirect.'id('butt').innerHTML='Go to '+ip;id('butt').disabled=false; IP: '+ip+'</font></b><br>Device is rebooting. Switch back to<br>the above WiFi network, and then<br>click the button below to redirect.'id('butt').onclick=function rd(){window.open('http://'+ip);}IP: '+ip+'</font></b><br>Device is rebooting. Switch back to<br>the above WiFi network, and then<br>click the button below to redirect.'clearInterval(tci);function rd(){window.open('http://'+ip);}IP: '+ip+'</font></b><br>Device is rebooting. Switch back to<br>the above WiFi network, and then<br>click the button below to redirect.'}learInterval(tci);function rd(){window.open('http://'+ip);}IP: '+ip+'</font></b><br>Device is rebooting. Switch back to<br>the above WiFi network, and then<br>click the button below to redirect.'}    Interval(tci);function rd(){window.open('http://'+ip);}IP: '+ip+'</font></b><br>Device is rebooting. Switch back to<br>the above WiFi network, and then<br>click the button below to redirect.'xhr.open('POST', 'jt', true); xhr.send();    'http://'+ip);}IP: '+ip+'</font></b><br>Device is rebooting. Switch back to<br>the above WiFi network, and then<br>click the button below to redirect.'}hr.open('POST', 'jt', true); xhr.send();    'http://'+ip);}IP: '+ip+'</font></b><br>Device is rebooting. Switch back to<br>the above WiFi network, and then<br>click the button below to redirect.'function sf() {, 'jt', true); xhr.send();    'http://'+ip);}IP: '+ip+'</font></b><br>Device is rebooting. Switch back to<br>the above WiFi network, and then<br>click the button below to redirect.'id('msg').innerHTML='';true); xhr.send();    'http://'+ip);}IP: '+ip+'</font></b><br>Device is rebooting. Switch back to<br>the above WiFi network, and then<br>click the button below to redirect.'var xhr=new XMLHttpRequest(); xhr.send();    'http://'+ip);}IP: '+ip+'</font></b><br>Device is rebooting. Switch back to<br>the above WiFi network, and then<br>click the button below to redirect.'xhr.onreadystatechange=function() {end();    'http://'+ip);}IP: '+ip+'</font></b><br>Device is rebooting. Switch back to<br>the above WiFi network, and then<br>click the button below to redirect.'if(xhr.readyState==4 && xhr.status==200) {   'http://'+ip);}IP: '+ip+'</font></b><br>Device is rebooting. Switch back to<br>the above WiFi network, and then<br>click the button below to redirect.'var jd=JSON.parse(xhr.responseText);200) {   'http://'+ip);}IP: '+ip+'</font></b><br>Device is rebooting. Switch back to<br>the above WiFi network, and then<br>click the button below to redirect.'if(jd.result==1) { tci=setInterval(tryConnect, 10000); return; }'+ip+'</font></b><br>Device is rebooting. Switch back to<br>the above WiFi network, and then<br>click the button below to redirect.'id('msg').innerHTML='<b><font color=red>Error code: '+jd.result+', item: '+jd.item+'</font></b>';booting. Switch back to<br>the above WiFi network, and then<br>click the button below to redirect.'id('butt').innerHTML='Submit';color=red>Error code: '+jd.result+', item: '+jd.item+'</font></b>';booting. Switch back to<br>the above WiFi network, and then<br>click the button below to redirect.'id('butt').disabled=false;id('ssid').disabled=false;id('pass').disabled=false;id('auth').disabled=false;. Switch back to<br>the above WiFi network, and then<br>click the button below to redirect.'}d('butt').disabled=false;id('ssid').disabled=false;id('pass').disabled=false;id('auth').disabled=false;. Switch back to<br>the above WiFi network, and then<br>click the button below to redirect.'}d('butt').disabled=false;id('ssid').disabled=false;id('pass').disabled=false;id('auth').disabled=false;. Switch back to<br>the above WiFi network, and then<br>click the button below to redirect.'var comm='cc?ssid='+encodeURIComponent(id('ssid').value)+'&pass='+encodeURIComponent(id('pass').value)+'&auth='+id('auth').value;bove WiFi network, and then<br>click the button below to redirect.'xhr.open('POST', comm, true); xhr.send();('ssid').value)+'&pass='+encodeURIComponent(id('pass').value)+'&auth='+id('auth').value;bove WiFi network, and then<br>click the button below to redirect.'id('butt').innerHTML='Connecting...';d();('ssid').value)+'&pass='+encodeURIComponent(id('pass').value)+'&auth='+id('auth').value;bove WiFi network, and then<br>click the button below to redirect.'id('butt').disabled=true;id('ssid').disabled=true;id('pass').disabled=true;id('auth').disabled=true;e)+'&auth='+id('auth').value;bove WiFi network, and then<br>click the button below to redirect.'}d('butt').disabled=true;id('ssid').disabled=true;id('pass').disabled=true;id('auth').disabled=true;e)+'&auth='+id('auth').value;bove WiFi network, and then<br>click the button below to redirect.'}d('butt').disabled=true;id('ssid').disabled=true;id('pass').disabled=true;id('auth').disabled=true;e)+'&auth='+id('auth').value;bove WiFi network, and then<br>click the button below to redirect.'function loadSSIDs() {ue;id('ssid').disabled=true;id('pass').disabled=true;id('auth').disabled=true;e)+'&auth='+id('auth').value;bove WiFi network, and then<br>click the button below to redirect.'var xhr=new XMLHttpRequest();ssid').disabled=true;id('pass').disabled=true;id('auth').disabled=true;e)+'&auth='+id('auth').value;bove WiFi network, and then<br>click the button below to redirect.'xhr.onreadystatechange=function() {.disabled=true;id('pass').disabled=true;id('auth').disabled=true;e)+'&auth='+id('auth').value;bove WiFi network, and then<br>click the button below to redirect.'if(xhr.readyState==4 && xhr.status==200) {ed=true;id('pass').disabled=true;id('auth').disabled=true;e)+'&auth='+id('auth').value;bove WiFi network, and then<br>click the button below to redirect.'var jd=JSON.parse(xhr.responseText), i;) {ed=true;id('pass').disabled=true;id('auth').disabled=true;e)+'&auth='+id('auth').value;bove WiFi network, and then<br>click the button below to redirect.'id('rd').deleteRow(0);responseText), i;) {ed=true;id('pass').disabled=true;id('auth').disabled=true;e)+'&auth='+id('auth').value;bove WiFi network, and then<br>click the button below to redirect.'for(i=0;i<jd.ssids.length;i++) {xt), i;) {ed=true;id('pass').disabled=true;id('auth').disabled=true;e)+'&auth='+id('auth').value;bove WiFi network, and then<br>click the button below to redirect.'var cell=id('rd').insertRow(-1).insertCell(0);rue;id('pass').disabled=true;id('auth').disabled=true;e)+'&auth='+id('auth').value;bove WiFi network, and then<br>click the button below to redirect.'cell.innerHTML ="<tr><td><input type='radio' name='ssids' id='rd" + i + "' value='" + jd.ssids[i] + "' onclick=sel(" + i + ")>" + jd.ssids[i] + "</td></tr>";br>click the button below to redirect.'}ell.innerHTML ="<tr><td><input type='radio' name='ssids' id='rd" + i + "' value='" + jd.ssids[i] + "' onclick=sel(" + i + ")>" + jd.ssids[i] + "</td></tr>";br>click the button below to redirect.'}ell.innerHTML ="<tr><td><input type='radio' name='ssids' id='rd" + i + "' value='" + jd.ssids[i] + "' onclick=sel(" + i + ")>" + jd.ssids[i] + "</td></tr>";br>click the button below to redirect.'};ll.innerHTML ="<tr><td><input type='radio' name='ssids' id='rd" + i + "' value='" + jd.ssids[i] + "' onclick=sel(" + i + ")>" + jd.ssids[i] + "</td></tr>";br>click the button below to redirect.'xhr.open('GET','js',true); xhr.send();radio' name='ssids' id='rd" + i + "' value='" + jd.ssids[i] + "' onclick=sel(" + i + ")>" + jd.ssids[i] + "</td></tr>";br>click the button below to redirect.'}hr.open('GET','js',true); xhr.send();radio' name='ssids' id='rd" + i + "' value='" + jd.ssids[i] + "' onclick=sel(" + i + ")>" + jd.ssids[i] + "</td></tr>";br>click the button below to redirect.'setTimeout(loadSSIDs, 1000);hr.send();radio' name='ssids' id='rd" + i + "' value='" + jd.ssids[i] + "' onclick=sel(" + i + ")>" + jd.ssids[i] + "</td></tr>";br>click the button below to redirect.'</script>t(loadSSIDs, 1000);hr.send();radio' name='ssids' id='rd" + i + "' value='" + jd.ssids[i] + "' onclick=sel(" + i + ")>" + jd.ssids[i] + "</td></tr>";br>click the button below to redirect.'</body>)";const char WWW_ASSETS["html_sta_home"] PROGMEM = R"(<head><title>OpenGarage</title><meta name='viewport' content='width=device-width, initial-scale=1'><link rel='stylesheet' href='http://code.jquery.com/mobile/1.3.1/jquery.mobile-1.3.1.min.css' type='text/css'><script src='http://code.jquery.com/jquery-1.9.1.min.js' type='text/javascript'></script>css' type='text/css'><script src='http://code.jquery.com/mobile/1.3.1/jquery.mobile-1.3.1.min.js' type='text/javascript'></script>></head> src='http://code.jquery.com/mobile/1.3.1/jquery.mobile-1.3.1.min.js' type='text/javascript'></script>><body>> src='http://code.jquery.com/mobile/1.3.1/jquery.mobile-1.3.1.min.js' type='text/javascript'></script>><div data-role='page' id='page_home'>obile/1.3.1/jquery.mobile-1.3.1.min.js' type='text/javascript'></script>><div data-role='header'><h3 id='head_name'>OG</h3></div>mobile-1.3.1.min.js' type='text/javascript'></script>><div data-role='content'>h3 id='head_name'>OG</h3></div>mobile-1.3.1.min.js' type='text/javascript'></script>><div data-role='fieldcontain'>='head_name'>OG</h3></div>mobile-1.3.1.min.js' type='text/javascript'></script>><table cellpadding='4'>ntain'>='head_name'>OG</h3></div>mobile-1.3.1.min.js' type='text/javascript'></script>><tr><td><b>Distance:</b></td><td><label id='lbl_dist'>-</label></td></tr>js' type='text/javascript'></script>><tr><td><b>Door State:</b></td><td><label id='lbl_status'>-</label></td></tr>type='text/javascript'></script>><tr><td><b>Read Count:</b></td><td><label id='lbl_beat'>-</label></td></tr>     e='text/javascript'></script>><tr><td><b>Device Key:</b></td><td><input type='password' size=16 maxlength=32 name='dkey' id='dkey'></td></tr><tr><td colspan=2><label id='msg'></label></td></tr>word' size=16 maxlength=32 name='dkey' id='dkey'></td></tr></table>colspan=2><label id='msg'></label></td></tr>word' size=16 maxlength=32 name='dkey' id='dkey'></td></tr><br />e>colspan=2><label id='msg'></label></td></tr>word' size=16 maxlength=32 name='dkey' id='dkey'></td></tr><div data-role='controlgroup' data-type='horizontal'>ord' size=16 maxlength=32 name='dkey' id='dkey'></td></tr><button data-theme='b' id='btn_click'>Button</button>  d' size=16 maxlength=32 name='dkey' id='dkey'></td></tr><button data-theme='b' id='btn_opts'>Options</button>  d' size=16 maxlength=32 name='dkey' id='dkey'></td></tr><button data-theme='b' id='btn_log'>Log</button>tton>  d' size=16 maxlength=32 name='dkey' id='dkey'></td></tr><button data-theme='c' id='btn_rbt'>Reboot</button>n>  d' size=16 maxlength=32 name='dkey' id='dkey'></td></tr></div>n data-theme='c' id='btn_rbt'>Reboot</button>n>  d' size=16 maxlength=32 name='dkey' id='dkey'></td></tr></div>n data-theme='c' id='btn_rbt'>Reboot</button>n>  d' size=16 maxlength=32 name='dkey' id='dkey'></td></tr></div>n data-theme='c' id='btn_rbt'>Reboot</button>n>  d' size=16 maxlength=32 name='dkey' id='dkey'></td></tr><div data-role='footer' data-theme='c'>oot</button>n>  d' size=16 maxlength=32 name='dkey' id='dkey'></td></tr><p>&nbsp; OpenGarage Firmware <label id='fwv'>-</label>&nbsp;<a href="update" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p></div>sp; OpenGarage Firmware <label id='fwv'>-</label>&nbsp;<a href="update" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p></div>sp; OpenGarage Firmware <label id='fwv'>-</label>&nbsp;<a href="update" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p><script>; OpenGarage Firmware <label id='fwv'>-</label>&nbsp;<a href="update" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>var si;>; OpenGarage Firmware <label id='fwv'>-</label>&nbsp;<a href="update" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>function clear_msg() {$('#msg').text('');}wv'>-</label>&nbsp;<a href="update" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>function show_msg(s,t,c) {msg').text('');}wv'>-</label>&nbsp;<a href="update" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>$('#msg').text(s).css('color',c);ext('');}wv'>-</label>&nbsp;<a href="update" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>if(t>0) setTimeout(clear_msg, t);ext('');}wv'>-</label>&nbsp;<a href="update" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>}f(t>0) setTimeout(clear_msg, t);ext('');}wv'>-</label>&nbsp;<a href="update" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>$('#btn_opts').click(function(e){window.open('vo');});>&nbsp;<a href="update" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>$('#btn_log').click(function(e){window.open('vl');});;>&nbsp;<a href="update" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>$('#btn_rbt').click(function(e){window.open('vl');});;>&nbsp;<a href="update" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>if(confirm('Reboot the device now?')){.open('vl');});;>&nbsp;<a href="update" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>var comm = 'cc?reboot=1&dkey='+($('#dkey').val());});;>&nbsp;<a href="update" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>clear_msg();cc?reboot=1&dkey='+($('#dkey').val());});;>&nbsp;<a href="update" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>$.getJSON(comm, function(jd) {+($('#dkey').val());});;>&nbsp;<a href="update" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>if(jd.result!=1) show_msg('Check device key and try again.',2000,'red');date" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>else {result!=1) show_msg('Check device key and try again.',2000,'red');date" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>show_msg('Rebooting. Please wait...',0,'green');try again.',2000,'red');date" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>clearInterval(si);g. Please wait...',0,'green');try again.',2000,'red');date" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>setTimeout(function(){location.reload(true);}, 10000);ain.',2000,'red');date" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>}etTimeout(function(){location.reload(true);}, 10000);ain.',2000,'red');date" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>});Timeout(function(){location.reload(true);}, 10000);ain.',2000,'red');date" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>});Timeout(function(){location.reload(true);}, 10000);ain.',2000,'red');date" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>});    out(function(){location.reload(true);}, 10000);ain.',2000,'red');date" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>$('#btn_click').click(function(e) {ad(true);}, 10000);ain.',2000,'red');date" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>var comm = 'cc?click=1&dkey='+($('#dkey').val());000);ain.',2000,'red');date" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>clear_msg();cc?click=1&dkey='+($('#dkey').val());000);ain.',2000,'red');date" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>$.getJSON(comm, function(jd) {($('#dkey').val());000);ain.',2000,'red');date" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>if(jd.result!=1) {nction(jd) {($('#dkey').val());000);ain.',2000,'red');date" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>show_msg('Check device key and try again.',2000,'red');in.',2000,'red');date" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>}how_msg('Check device key and try again.',2000,'red');in.',2000,'red');date" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>});w_msg('Check device key and try again.',2000,'red');in.',2000,'red');date" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>});w_msg('Check device key and try again.',2000,'red');in.',2000,'red');date" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>$(document).ready(function() { try again.',2000,'red');in.',2000,'red');date" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>show(); si=setInterval('show()', 3000);n.',2000,'red');in.',2000,'red');date" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>});w(); si=setInterval('show()', 3000);n.',2000,'red');in.',2000,'red');date" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>function show() {erval('show()', 3000);n.',2000,'red');in.',2000,'red');date" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>$.getJSON('jc', function(jd) {', 3000);n.',2000,'red');in.',2000,'red');date" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>$('#fwv').text('v'+(jd.fwv/100>>0)+'.'+(jd.fwv/10%10>>0)+'.'+(jd.fwv%10>>0)); target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>$('#lbl_dist').text(''+jd.dist+' (cm)');jd.fwv/10%10>>0)+'.'+(jd.fwv%10>>0)); target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>$('#lbl_status').text(jd.door?'OPEN':'closed').css('color', jd.door?'red':'black');t='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>$('#lbl_beat').text(jd.rcnt);?'OPEN':'closed').css('color', jd.door?'red':'black');t='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>$('#head_name').text(jd.name);'OPEN':'closed').css('color', jd.door?'red':'black');t='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>});#head_name').text(jd.name);'OPEN':'closed').css('color', jd.door?'red':'black');t='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>});#head_name').text(jd.name);'OPEN':'closed').css('color', jd.door?'red':'black');t='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p></script>name').text(jd.name);'OPEN':'closed').css('color', jd.door?'red':'black');t='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p></body>)";const char WWW_ASSETS["html_sta_options"] PROGMEM = R"(<!DOCTYPE html><head>YPE html><title>OpenGarage Options</title><meta name='viewport' content='width=device-width, initial-scale=1'><link rel='stylesheet' href='http://code.jquery.com/mobile/1.3.1/jquery.mobile-1.3.1.min.css' type='text/css'><script src='http://code.jquery.com/jquery-1.9.1.min.js' type='text/javascript'></script>css' type='text/css'><script src='http://code.jquery.com/mobile/1.3.1/jquery.mobile-1.3.1.min.js' type='text/javascript'></script>></head> src='http://code.jquery.com/mobile/1.3.1/jquery.mobile-1.3.1.min.js' type='text/javascript'></script>><body>> src='http://code.jquery.com/mobile/1.3.1/jquery.mobile-1.3.1.min.js' type='text/javascript'></script>><div data-role='page' id='page_opts'>obile/1.3.1/jquery.mobile-1.3.1.min.js' type='text/javascript'></script>><div data-role='header'><h3>Edit Options</h3></div>    .mobile-1.3.1.min.js' type='text/javascript'></script>><div data-role='content'>   Edit Options</h3></div>    .mobile-1.3.1.min.js' type='text/javascript'></script>><table cellpadding=2>nt'>   Edit Options</h3></div>    .mobile-1.3.1.min.js' type='text/javascript'></script>><tr><td><b>Accessibility:</b></td><td>ns</h3></div>    .mobile-1.3.1.min.js' type='text/javascript'></script>><select name='acc' id='acc' data-mini='true'></div>    .mobile-1.3.1.min.js' type='text/javascript'></script>><option value=0>Direct IP Only</option>true'></div>    .mobile-1.3.1.min.js' type='text/javascript'></script>><option value=1>Direct IP + Cloud</option>                  le-1.3.1.min.js' type='text/javascript'></script>><option value=2>Cloud Only</option>option>                  le-1.3.1.min.js' type='text/javascript'></script>></select></td></tr>      y</option>option>                  le-1.3.1.min.js' type='text/javascript'></script>><tr><td><b>Cloud Token:</b></td><td><input type='text' size=24 maxlength=32 id='auth' data-mini='true' value='-'></td></tr><tr><td><b>Mount Type:</b></td><td>><input type='text' size=24 maxlength=32 id='auth' data-mini='true' value='-'></td></tr><select name='mnt' id='mnt' data-mini='true'>pe='text' size=24 maxlength=32 id='auth' data-mini='true' value='-'></td></tr><option value=0>Ceiling Mount</option>'true'>pe='text' size=24 maxlength=32 id='auth' data-mini='true' value='-'></td></tr><option value=1>Side Mount</option>on>'true'>pe='text' size=24 maxlength=32 id='auth' data-mini='true' value='-'></td></tr></select></td></tr>  Mount</option>on>'true'>pe='text' size=24 maxlength=32 id='auth' data-mini='true' value='-'></td></tr><tr><td><b>Threshold (cm): </b></td><td><input type='text' size=4 maxlength=4 id='dth' data-mini='true' value=1></td></tr>><tr><td><b>Read Interval (s):</b></td><td><input type='text' size=3 maxlength=3 id='riv' data-mini='true' value=1></td></tr><tr><td><b>Sound Alarm:</b></td><td>d><td><input type='text' size=3 maxlength=3 id='riv' data-mini='true' value=1></td></tr><select name='alm' id='alm' data-mini='true'>      pe='text' size=3 maxlength=3 id='riv' data-mini='true' value=1></td></tr><option value=0>Disabled</option>mini='true'>      pe='text' size=3 maxlength=3 id='riv' data-mini='true' value=1></td></tr><option value=1>5 seconds</option>                  e='text' size=3 maxlength=3 id='riv' data-mini='true' value=1></td></tr><option value=2>10 seconds</option>                 e='text' size=3 maxlength=3 id='riv' data-mini='true' value=1></td></tr></select></td></tr>seconds</option>                 e='text' size=3 maxlength=3 id='riv' data-mini='true' value=1></td></tr><tr><td><b>Device Name:</b></td><td><input type='text' size=20 maxlength=32 id='name' data-mini='true' value='-'></td></tr>><tr><td><b>HTTP Port:</b></td><td><input type='text' size=5 maxlength=5 id='htp' value=1 data-mini='true'></td></tr>d></tr>><tr><td><b>Device Key:</b></td><td><input type='password' size=24 maxlength=32 id='dkey' data-mini='true'></td></tr>d></tr>><tr><td colspan=2><p id='msg'></p></td></tr>pe='password' size=24 maxlength=32 id='dkey' data-mini='true'></td></tr>d></tr>><tr><td colspan=2><input type='checkbox' data-mini='true' id='cb_key'><label for='cb_key'>Change Device Key</label></td></tr><tr><td><b>New Key:</b></td><td><input type='password' size=24 maxlength=32 id='nkey' data-mini='true' disabled></td></tr>tr><tr><td><b>Confirm:</b></td><td><input type='password' size=24 maxlength=32 id='ckey' data-mini='true' disabled></td></tr>  ></table><b>Confirm:</b></td><td><input type='password' size=24 maxlength=32 id='ckey' data-mini='true' disabled></td></tr>  ><div data-role='controlgroup' data-type='horizontal'>' size=24 maxlength=32 id='ckey' data-mini='true' disabled></td></tr>  ><a href='#' data-role='button' data-inline='true' data-theme='a' id='btn_cancel'>Cancel</a>mini='true' disabled></td></tr>  ><a href='#' data-role='button' data-inline='true' data-theme='b' id='btn_submit'>Submit</a>      true' disabled></td></tr>  ></div>f='#' data-role='button' data-inline='true' data-theme='b' id='btn_submit'>Submit</a>      true' disabled></td></tr>  ></div>f='#' data-role='button' data-inline='true' data-theme='b' id='btn_submit'>Submit</a>      true' disabled></td></tr>  ><div data-role='footer' data-theme='c'>ine='true' data-theme='b' id='btn_submit'>Submit</a>      true' disabled></td></tr>  ><p>&nbsp; OpenGarage Firmware <label id='fwv'>-</label>&nbsp;<a href="update" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p></div> p; OpenGarage Firmware <label id='fwv'>-</label>&nbsp;<a href="update" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p></div> p; OpenGarage Firmware <label id='fwv'>-</label>&nbsp;<a href="update" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p><script>; OpenGarage Firmware <label id='fwv'>-</label>&nbsp;<a href="update" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>//#msg acts as a response container for events-</label>&nbsp;<a href="update" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>function clear_msg() {nse container for events-</label>&nbsp;<a href="update" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>$('#msg').text('');) {nse container for events-</label>&nbsp;<a href="update" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>}  #msg').text('');) {nse container for events-</label>&nbsp;<a href="update" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>function show_msg(s) {nse container for events-</label>&nbsp;<a href="update" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>$('#msg').text(s).css('color','red');or events-</label>&nbsp;<a href="update" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>setTimeout(clear_msg, 2000);','red');or events-</label>&nbsp;<a href="update" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>}etTimeout(clear_msg, 2000);','red');or events-</label>&nbsp;<a href="update" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>// enable/disable the change key textboxes when the 'change key'href="update" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>// checkbox is changedchange key textboxes when the 'change key'href="update" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>$('#cb_key').click(function(e){y textboxes when the 'change key'href="update" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>$('#nkey').textinput($(this).is(':checked')?'enable':'disable');href="update" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>$('#ckey').textinput($(this).is(':checked')?'enable':'disable');href="update" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>});#ckey').textinput($(this).is(':checked')?'enable':'disable');href="update" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>$('#btn_cancel').click(function(e){hecked')?'enable':'disable');href="update" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>e.preventDefault(); close();ion(e){hecked')?'enable':'disable');href="update" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>});reventDefault(); close();ion(e){hecked')?'enable':'disable');href="update" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>});reventDefault(); close();ion(e){hecked')?'enable':'disable');href="update" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>$('#btn_submit').click(function(e){hecked')?'enable':'disable');href="update" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>e.preventDefault();ick(function(e){hecked')?'enable':'disable');href="update" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>if(confirm('Submit changes?')) {e){hecked')?'enable':'disable');href="update" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>var comm='co?dkey='+encodeURIComponent($('#dkey').val());able');href="update" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>comm+='&acc='+$('#acc').val();omponent($('#dkey').val());able');href="update" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>comm+='&mnt='+$('#mnt').val();omponent($('#dkey').val());able');href="update" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>comm+='&dth='+$('#dth').val();omponent($('#dkey').val());able');href="update" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>comm+='&riv='+$('#riv').val();omponent($('#dkey').val());able');href="update" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>comm+='&alm='+$('#alm').val();omponent($('#dkey').val());able');href="update" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>comm+='&htp='+$('#htp').val();omponent($('#dkey').val());able');href="update" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>comm+='&name='+encodeURIComponent($('#name').val());l());able');href="update" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>comm+='&auth='+encodeURIComponent($('#auth').val());l());able');href="update" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>if($('#cb_key').is(':checked')) {($('#auth').val());l());able');href="update" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>if(!$('#nkey').val()) {ecked')) {($('#auth').val());l());able');href="update" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>if(!confirm('New device key is empty. Are you sure?')) return;);href="update" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>}f(!confirm('New device key is empty. Are you sure?')) return;);href="update" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>comm+='&nkey='+encodeURIComponent($('#nkey').val());)) return;);href="update" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>comm+='&ckey='+encodeURIComponent($('#ckey').val());)) return;);href="update" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>}omm+='&ckey='+encodeURIComponent($('#ckey').val());)) return;);href="update" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>$.getJSON(comm, function(jd) {ent($('#ckey').val());)) return;);href="update" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>if(jd.result!=1) {nction(jd) {ent($('#ckey').val());)) return;);href="update" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>if(jd.result==2) show_msg('Check device key and try again.');;);href="update" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>else show_msg('Error code: '+jd.result+', item: '+jd.item););;);href="update" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>} else {w_msg('Error code: '+jd.result+', item: '+jd.item););;);href="update" target='_blank' data-role='button' data-inline=true data-mini=true>Update</a></p>$('#msg').html('<font color=green>Options are successfully saved. Note that<br>changes to some options may require a reboot.</font>');-mini=true>Update</a></p>$('#btn_submit').font color=green>Options are successfully saved. Note that<br>changes to some options may require a reboot.</font>');-mini=true>Update</a></p>setTimeout(close, 4000);lor=green>Options are successfully saved. Note that<br>changes to some options may require a reboot.</font>');-mini=true>Update</a></p>}etTimeout(close, 4000);lor=green>Options are successfully saved. Note that<br>changes to some options may require a reboot.</font>');-mini=true>Update</a></p>});Timeout(close, 4000);lor=green>Options are successfully saved. Note that<br>changes to some options may require a reboot.</font>');-mini=true>Update</a></p>});Timeout(close, 4000);lor=green>Options are successfully saved. Note that<br>changes to some options may require a reboot.</font>');-mini=true>Update</a></p>});Timeout(close, 4000);lor=green>Options are successfully saved. Note that<br>changes to some options may require a reboot.</font>');-mini=true>Update</a></p>$(document).ready(function() {een>Options are successfully saved. Note that<br>changes to some options may require a reboot.</font>');-mini=true>Update</a></p>$.getJSON('jo', function(jd) {een>Options are successfully saved. Note that<br>changes to some options may require a reboot.</font>');-mini=true>Update</a></p>$('#fwv').text('v'+(jd.fwv/100>>0)+'.'+(jd.fwv/10%10>>0)+'.'+(jd.fwv%10>>0));r>changes to some options may require a reboot.</font>');-mini=true>Update</a></p>$('#acc').val(jd.acc).selectmenu('refresh');wv/10%10>>0)+'.'+(jd.fwv%10>>0));r>changes to some options may require a reboot.</font>');-mini=true>Update</a></p>$('#alm').val(jd.alm).selectmenu('refresh');wv/10%10>>0)+'.'+(jd.fwv%10>>0));r>changes to some options may require a reboot.</font>');-mini=true>Update</a></p>$('#mnt').val(jd.mnt).selectmenu('refresh');wv/10%10>>0)+'.'+(jd.fwv%10>>0));r>changes to some options may require a reboot.</font>');-mini=true>Update</a></p>$('#dth').val(jd.dth);selectmenu('refresh');wv/10%10>>0)+'.'+(jd.fwv%10>>0));r>changes to some options may require a reboot.</font>');-mini=true>Update</a></p>$('#riv').val(jd.riv);selectmenu('refresh');wv/10%10>>0)+'.'+(jd.fwv%10>>0));r>changes to some options may require a reboot.</font>');-mini=true>Update</a></p>$('#htp').val(jd.htp);selectmenu('refresh');wv/10%10>>0)+'.'+(jd.fwv%10>>0));r>changes to some options may require a reboot.</font>');-mini=true>Update</a></p>$('#name').val(jd.name);lectmenu('refresh');wv/10%10>>0)+'.'+(jd.fwv%10>>0));r>changes to some options may require a reboot.</font>');-mini=true>Update</a></p>$('#auth').val(jd.auth);lectmenu('refresh');wv/10%10>>0)+'.'+(jd.fwv%10>>0));r>changes to some options may require a reboot.</font>');-mini=true>Update</a></p>});#auth').val(jd.auth);lectmenu('refresh');wv/10%10>>0)+'.'+(jd.fwv%10>>0));r>changes to some options may require a reboot.</font>');-mini=true>Update</a></p>});#auth').val(jd.auth);lectmenu('refresh');wv/10%10>>0)+'.'+(jd.fwv%10>>0));r>changes to some options may require a reboot.</font>');-mini=true>Update</a></p></script>).val(jd.auth);lectmenu('refresh');wv/10%10>>0)+'.'+(jd.fwv%10>>0));r>changes to some options may require a reboot.</font>');-mini=true>Update</a></p></body>)";const char WWW_ASSETS["html_sta_update"] PROGMEM = R"(<head><title>OpenGarage</title><meta name='viewport' content='width=device-width, initial-scale=1'><link rel='stylesheet' href='http://code.jquery.com/mobile/1.3.1/jquery.mobile-1.3.1.min.css' type='text/css'><script src='http://code.jquery.com/jquery-1.9.1.min.js' type='text/javascript'></script>css' type='text/css'><script src='http://code.jquery.com/mobile/1.3.1/jquery.mobile-1.3.1.min.js' type='text/javascript'></script>></head> src='http://code.jquery.com/mobile/1.3.1/jquery.mobile-1.3.1.min.js' type='text/javascript'></script>><body>> src='http://code.jquery.com/mobile/1.3.1/jquery.mobile-1.3.1.min.js' type='text/javascript'></script>><div data-role='page' id='page_update'>ile/1.3.1/jquery.mobile-1.3.1.min.js' type='text/javascript'></script>><div data-role='header'><h3>OpenGarage Firmware Update</h3></div>3.1.min.js' type='text/javascript'></script>><div data-role='content'>h3>OpenGarage Firmware Update</h3></div>3.1.min.js' type='text/javascript'></script>><form method='POST' action='/update' id='fm' enctype='multipart/form-data'>' type='text/javascript'></script>><table cellspacing=4>ction='/update' id='fm' enctype='multipart/form-data'>' type='text/javascript'></script>><tr><td><input type='file' name='file' accept='.bin' id='file'></td></tr>'>' type='text/javascript'></script>><tr><td><b>Device key: </b><input type='password' name='dkey' size=16 maxlength=16 id='dkey'></td></tr>cript>><tr><td><label id='msg'></label></td></tr>ssword' name='dkey' size=16 maxlength=16 id='dkey'></td></tr>cript>></table><label id='msg'></label></td></tr>ssword' name='dkey' size=16 maxlength=16 id='dkey'></td></tr>cript>><a href='#' data-role='button' data-inline='true' data-theme='a' id='btn_cancel'>Cancel</a>'></td></tr>cript>><a href='#' data-role='button' data-inline='true' data-theme='b' id='btn_submit'>Submit</a>'></td></tr>cript>></form>='#' data-role='button' data-inline='true' data-theme='b' id='btn_submit'>Submit</a>'></td></tr>cript>></div>>='#' data-role='button' data-inline='true' data-theme='b' id='btn_submit'>Submit</a>'></td></tr>cript>></div>>='#' data-role='button' data-inline='true' data-theme='b' id='btn_submit'>Submit</a>'></td></tr>cript>><script>'#' data-role='button' data-inline='true' data-theme='b' id='btn_submit'>Submit</a>'></td></tr>cript>>function id(s) {return document.getElementById(s);}ata-theme='b' id='btn_submit'>Submit</a>'></td></tr>cript>>function clear_msg() {id('msg').innerHTML='';}(s);}ata-theme='b' id='btn_submit'>Submit</a>'></td></tr>cript>>function show_msg(s,t,c) {msg').innerHTML='';}(s);}ata-theme='b' id='btn_submit'>Submit</a>'></td></tr>cript>>id('msg').innerHTML=s.fontcolor(c);erHTML='';}(s);}ata-theme='b' id='btn_submit'>Submit</a>'></td></tr>cript>>if(t>0) setTimeout(clear_msg, t););erHTML='';}(s);}ata-theme='b' id='btn_submit'>Submit</a>'></td></tr>cript>>}  t>0) setTimeout(clear_msg, t););erHTML='';}(s);}ata-theme='b' id='btn_submit'>Submit</a>'></td></tr>cript>>$('#btn_cancel').click(function(e){erHTML='';}(s);}ata-theme='b' id='btn_submit'>Submit</a>'></td></tr>cript>>e.preventDefault(); close();ion(e){erHTML='';}(s);}ata-theme='b' id='btn_submit'>Submit</a>'></td></tr>cript>>});reventDefault(); close();ion(e){erHTML='';}(s);}ata-theme='b' id='btn_submit'>Submit</a>'></td></tr>cript>>$('#btn_submit').click(function(e){erHTML='';}(s);}ata-theme='b' id='btn_submit'>Submit</a>'></td></tr>cript>>var files= id('file').files;ion(e){erHTML='';}(s);}ata-theme='b' id='btn_submit'>Submit</a>'></td></tr>cript>>if(files.length==0) {show_msg('Please select a file.',2000,'red'); return;}bmit'>Submit</a>'></td></tr>cript>>if(id('dkey').value=='') {msg('Please select a file.',2000,'red'); return;}bmit'>Submit</a>'></td></tr>cript>>if(!confirm('You did not input a device key. Are you sure?')) return;turn;}bmit'>Submit</a>'></td></tr>cript>>}f(!confirm('You did not input a device key. Are you sure?')) return;turn;}bmit'>Submit</a>'></td></tr>cript>>var btn = id('btn_submit');put a device key. Are you sure?')) return;turn;}bmit'>Submit</a>'></td></tr>cript>>show_msg('Uploading. Please wait...',0,'green'); you sure?')) return;turn;}bmit'>Submit</a>'></td></tr>cript>>var fd = new FormData();ase wait...',0,'green'); you sure?')) return;turn;}bmit'>Submit</a>'></td></tr>cript>>var file = files[0];a();ase wait...',0,'green'); you sure?')) return;turn;}bmit'>Submit</a>'></td></tr>cript>>fd.append('file', file, file.name);',0,'green'); you sure?')) return;turn;}bmit'>Submit</a>'></td></tr>cript>>fd.append('dkey', id('dkey').value);,0,'green'); you sure?')) return;turn;}bmit'>Submit</a>'></td></tr>cript>>var xhr = new XMLHttpRequest();lue);,0,'green'); you sure?')) return;turn;}bmit'>Submit</a>'></td></tr>cript>>xhr.onreadystatechange = function() {0,'green'); you sure?')) return;turn;}bmit'>Submit</a>'></td></tr>cript>>if(xhr.readyState==4 && xhr.status==200) {een'); you sure?')) return;turn;}bmit'>Submit</a>'></td></tr>cript>>var jd=JSON.parse(xhr.responseText);200) {een'); you sure?')) return;turn;}bmit'>Submit</a>'></td></tr>cript>>if(jd.result==1) {xhr.responseText);200) {een'); you sure?')) return;turn;}bmit'>Submit</a>'></td></tr>cript>>show_msg('Update is successful. Rebooting. Please wait...',0,'green');urn;}bmit'>Submit</a>'></td></tr>cript>>setTimeout(close, 10000);ssful. Rebooting. Please wait...',0,'green');urn;}bmit'>Submit</a>'></td></tr>cript>>} else if (jd.result==2) {sful. Rebooting. Please wait...',0,'green');urn;}bmit'>Submit</a>'></td></tr>cript>>show_msg('Check device key and try again.', 0, 'red');...',0,'green');urn;}bmit'>Submit</a>'></td></tr>cript>>} else {('Check device key and try again.', 0, 'red');...',0,'green');urn;}bmit'>Submit</a>'></td></tr>cript>>show_msg('Update failed.',0,'red');again.', 0, 'red');...',0,'green');urn;}bmit'>Submit</a>'></td></tr>cript>>}how_msg('Update failed.',0,'red');again.', 0, 'red');...',0,'green');urn;}bmit'>Submit</a>'></td></tr>cript>>}how_msg('Update failed.',0,'red');again.', 0, 'red');...',0,'green');urn;}bmit'>Submit</a>'></td></tr>cript>>};ow_msg('Update failed.',0,'red');again.', 0, 'red');...',0,'green');urn;}bmit'>Submit</a>'></td></tr>cript>>xhr.open('POST', 'update', true););again.', 0, 'red');...',0,'green');urn;}bmit'>Submit</a>'></td></tr>cript>>xhr.send(fd);T', 'update', true););again.', 0, 'red');...',0,'green');urn;}bmit'>Submit</a>'></td></tr>cript>>});.send(fd);T', 'update', true););again.', 0, 'red');...',0,'green');urn;}bmit'>Submit</a>'></td></tr>cript>></script>fd);T', 'update', true););again.', 0, 'red');...',0,'green');urn;}bmit'>Submit</a>'></td></tr>cript>></body>)";const char WWW_ASSETS["html_sta_logs"] PROGMEM = R"(<head><title>OpenGarage</title><meta name='viewport' content='width=device-width, initial-scale=1'><link rel='stylesheet' href='http://code.jquery.com/mobile/1.3.1/jquery.mobile-1.3.1.min.css' type='text/css'><script src='http://code.jquery.com/jquery-1.9.1.min.js' type='text/javascript'></script>css' type='text/css'><script src='http://code.jquery.com/mobile/1.3.1/jquery.mobile-1.3.1.min.js' type='text/javascript'></script>></head> src='http://code.jquery.com/mobile/1.3.1/jquery.mobile-1.3.1.min.js' type='text/javascript'></script>><body>> src='http://code.jquery.com/mobile/1.3.1/jquery.mobile-1.3.1.min.js' type='text/javascript'></script>><div data-role='page' id='page_log'>mobile/1.3.1/jquery.mobile-1.3.1.min.js' type='text/javascript'></script>><div data-role='header'><h3><label id='lbl_name'></label> Log</h3></div>     type='text/javascript'></script>><div data-role='content'>h3><label id='lbl_name'></label> Log</h3></div>     type='text/javascript'></script>><p>Below are the most recent <label id='lbl_nr'></label> records</p>div>     type='text/javascript'></script>><p>Current time is <label id='lbl_time'></label></p>bel> records</p>div>     type='text/javascript'></script>><div data-role='fieldcontain'>lbl_time'></label></p>bel> records</p>div>     type='text/javascript'></script>><table id='tab_log' border='1' cellpadding='4' style='border-collapse:collapse;'><tr><td align='center'><b>Time Stamp</b></td><td align='center'><b>Status</b></td><td align='center'><b>D (cm)</b></td></tr></table></div> id='tab_log' border='1' cellpadding='4' style='border-collapse:collapse;'><tr><td align='center'><b>Time Stamp</b></td><td align='center'><b>Status</b></td><td align='center'><b>D (cm)</b></td></tr></table><div data-role="controlgroup" data-type="horizontal">'border-collapse:collapse;'><tr><td align='center'><b>Time Stamp</b></td><td align='center'><b>Status</b></td><td align='center'><b>D (cm)</b></td></tr></table><button data-theme="b" id="btn_close">Close</button>>'border-collapse:collapse;'><tr><td align='center'><b>Time Stamp</b></td><td align='center'><b>Status</b></td><td align='center'><b>D (cm)</b></td></tr></table></div>n data-theme="b" id="btn_close">Close</button>>'border-collapse:collapse;'><tr><td align='center'><b>Time Stamp</b></td><td align='center'><b>Status</b></td><td align='center'><b>D (cm)</b></td></tr></table></div>n data-theme="b" id="btn_close">Close</button>>'border-collapse:collapse;'><tr><td align='center'><b>Time Stamp</b></td><td align='center'><b>Status</b></td><td align='center'><b>D (cm)</b></td></tr></table></div>n data-theme="b" id="btn_close">Close</button>>'border-collapse:collapse;'><tr><td align='center'><b>Time Stamp</b></td><td align='center'><b>Status</b></td><td align='center'><b>D (cm)</b></td></tr></table><script>data-theme="b" id="btn_close">Close</button>>'border-collapse:collapse;'><tr><td align='center'><b>Time Stamp</b></td><td align='center'><b>Status</b></td><td align='center'><b>D (cm)</b></td></tr></table>var curr_time = 0;="b" id="btn_close">Close</button>>'border-collapse:collapse;'><tr><td align='center'><b>Time Stamp</b></td><td align='center'><b>Status</b></td><td align='center'><b>D (cm)</b></td></tr></table>var date = new Date(); id="btn_close">Close</button>>'border-collapse:collapse;'><tr><td align='center'><b>Time Stamp</b></td><td align='center'><b>Status</b></td><td align='center'><b>D (cm)</b></td></tr></table>$("#btn_close").click(function(){close();});/button>>'border-collapse:collapse;'><tr><td align='center'><b>Time Stamp</b></td><td align='center'><b>Status</b></td><td align='center'><b>D (cm)</b></td></tr></table>$(document).ready(function(){n(){close();});/button>>'border-collapse:collapse;'><tr><td align='center'><b>Time Stamp</b></td><td align='center'><b>Status</b></td><td align='center'><b>D (cm)</b></td></tr></table>show_log();.ready(function(){n(){close();});/button>>'border-collapse:collapse;'><tr><td align='center'><b>Time Stamp</b></td><td align='center'><b>Status</b></td><td align='center'><b>D (cm)</b></td></tr></table>setInterval(show_time, 1000);n(){close();});/button>>'border-collapse:collapse;'><tr><td align='center'><b>Time Stamp</b></td><td align='center'><b>Status</b></td><td align='center'><b>D (cm)</b></td></tr></table>});Interval(show_time, 1000);n(){close();});/button>>'border-collapse:collapse;'><tr><td align='center'><b>Time Stamp</b></td><td align='center'><b>Status</b></td><td align='center'><b>D (cm)</b></td></tr></table>function show_time() { 1000);n(){close();});/button>>'border-collapse:collapse;'><tr><td align='center'><b>Time Stamp</b></td><td align='center'><b>Status</b></td><td align='center'><b>D (cm)</b></td></tr></table>curr_time ++;_time() { 1000);n(){close();});/button>>'border-collapse:collapse;'><tr><td align='center'><b>Time Stamp</b></td><td align='center'><b>Status</b></td><td align='center'><b>D (cm)</b></td></tr></table>date.setTime(curr_time*1000);n(){close();});/button>>'border-collapse:collapse;'><tr><td align='center'><b>Time Stamp</b></td><td align='center'><b>Status</b></td><td align='center'><b>D (cm)</b></td></tr></table>$('#lbl_time').text(date.toLocaleString());;/button>>'border-collapse:collapse;'><tr><td align='center'><b>Time Stamp</b></td><td align='center'><b>Status</b></td><td align='center'><b>D (cm)</b></td></tr></table>}('#lbl_time').text(date.toLocaleString());;/button>>'border-collapse:collapse;'><tr><td align='center'><b>Time Stamp</b></td><td align='center'><b>Status</b></td><td align='center'><b>D (cm)</b></td></tr></table>function show_log() {ate.toLocaleString());;/button>>'border-collapse:collapse;'><tr><td align='center'><b>Time Stamp</b></td><td align='center'><b>Status</b></td><td align='center'><b>D (cm)</b></td></tr></table>$.getJSON('jl', function(jd) {aleString());;/button>>'border-collapse:collapse;'><tr><td align='center'><b>Time Stamp</b></td><td align='center'><b>Status</b></td><td align='center'><b>D (cm)</b></td></tr></table>$('#lbl_name').text(jd.name);{aleString());;/button>>'border-collapse:collapse;'><tr><td align='center'><b>Time Stamp</b></td><td align='center'><b>Status</b></td><td align='center'><b>D (cm)</b></td></tr></table>curr_time = jd.time;jd.name);{aleString());;/button>>'border-collapse:collapse;'><tr><td align='center'><b>Time Stamp</b></td><td align='center'><b>Status</b></td><td align='center'><b>D (cm)</b></td></tr></table>$('#tab_log').find('tr:gt(0)').remove();));;/button>>'border-collapse:collapse;'><tr><td align='center'><b>Time Stamp</b></td><td align='center'><b>Status</b></td><td align='center'><b>D (cm)</b></td></tr></table>var logs=jd.logs;d('tr:gt(0)').remove();));;/button>>'border-collapse:collapse;'><tr><td align='center'><b>Time Stamp</b></td><td align='center'><b>Status</b></td><td align='center'><b>D (cm)</b></td></tr></table>logs.sort(function(a,b){return b[0]-a[0];});/button>>'border-collapse:collapse;'><tr><td align='center'><b>Time Stamp</b></td><td align='center'><b>Status</b></td><td align='center'><b>D (cm)</b></td></tr></table>$('#lbl_nr').text(logs.length);b[0]-a[0];});/button>>'border-collapse:collapse;'><tr><td align='center'><b>Time Stamp</b></td><td align='center'><b>Status</b></td><td align='center'><b>D (cm)</b></td></tr></table>var ldate = new Date();length);b[0]-a[0];});/button>>'border-collapse:collapse;'><tr><td align='center'><b>Time Stamp</b></td><td align='center'><b>Status</b></td><td align='center'><b>D (cm)</b></td></tr></table>for(var i=0;i<logs.length;i++) {[0]-a[0];});/button>>'border-collapse:collapse;'><tr><td align='center'><b>Time Stamp</b></td><td align='center'><b>Status</b></td><td align='center'><b>D (cm)</b></td></tr></table>ldate.setTime(logs[i][0]*1000);{[0]-a[0];});/button>>'border-collapse:collapse;'><tr><td align='center'><b>Time Stamp</b></td><td align='center'><b>Status</b></td><td align='center'><b>D (cm)</b></td></tr></table>var r='<tr><td align="center">'+ldate.toLocaleString()+'</td><td align="center">'+(logs[i][1]?'OPEN':'closed')+'</td><td align="center">'+logs[i][2]+'</td></tr>';><td align='center'><b>D (cm)</b></td></tr></table>$('#tab_log').append(r);nter">'+ldate.toLocaleString()+'</td><td align="center">'+(logs[i][1]?'OPEN':'closed')+'</td><td align="center">'+logs[i][2]+'</td></tr>';><td align='center'><b>D (cm)</b></td></tr></table>}('#tab_log').append(r);nter">'+ldate.toLocaleString()+'</td><td align="center">'+(logs[i][1]?'OPEN':'closed')+'</td><td align="center">'+logs[i][2]+'</td></tr>';><td align='center'><b>D (cm)</b></td></tr></table>});#tab_log').append(r);nter">'+ldate.toLocaleString()+'</td><td align="center">'+(logs[i][1]?'OPEN':'closed')+'</td><td align="center">'+logs[i][2]+'</td></tr>';><td align='center'><b>D (cm)</b></td></tr></table>setTimeout(show_log, 10000);">'+ldate.toLocaleString()+'</td><td align="center">'+(logs[i][1]?'OPEN':'closed')+'</td><td align="center">'+logs[i][2]+'</td></tr>';><td align='center'><b>D (cm)</b></td></tr></table>}etTimeout(show_log, 10000);">'+ldate.toLocaleString()+'</td><td align="center">'+(logs[i][1]?'OPEN':'closed')+'</td><td align="center">'+logs[i][2]+'</td></tr>';><td align='center'><b>D (cm)</b></td></tr></table></script>t(show_log, 10000);">'+ldate.toLocaleString()+'</td><td align="center">'+(logs[i][1]?'OPEN':'closed')+'</td><td align="center">'+logs[i][2]+'</td></tr>';><td align='center'><b>D (cm)</b></td></tr></table></body>)";const char WWW_ASSETS["portal"] PROGMEM = R"()";
