+++ OpenGarage/www_assets/publish_assets.cpp
@@ -1,0 +1,90 @@
+// takes an html file input and viewname and creates a
+// file with a PROGMEM statement that contains the
+// original html
+
+#include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
+
+int main(int argc, char* argv[])
+{
+  // check if we were provided enough arguments, if not gripe
+  if(argc<3) {
+    printf("Usage: %s inputHtmlFile.html viewname\n", argv[0]);
+    return 0;
+  }
+
+  char oname[] = "../www_assets.h";
+
+  // open a pointer to the input file
+  FILE *fp = fopen(argv[1], "rb");
+  // and if we can't, toss a wobbly
+  if(!fp) {
+    printf("Can't open input file %s\n", argv[1]);
+    return 0;
+  }
+
+  // open a pointer to the output file
+  FILE *op = fopen(oname, "ab");
+  // and if we can't, toss a different wobbly
+  if(!op) {
+    printf("Can't open output file %s\n", oname);
+    return 0;
+  }
+
+  char in[1000];
+  char out[1000];
+  int size;
+  char c;
+  int i;
+
+  // start by writing our 'header' for this view to the file
+  fprintf(op, "const char WWW_ASSETS[\"%s\"] PROGMEM = R\"(", argv[2]);
+  bool in_body = false;
+  
+  while(!feof(fp)) {
+    fgets(in, sizeof(in), fp);
+    size = strlen(in);
+
+    c = in[size-1];
+    if(c=='\r' || c=='\n') size--;
+    c = in[size-1];
+    if(c=='\r' || c=='\n') size--;
+    c = in[size-1];
+    if(c=='\r' || c=='\n') size--;
+    
+    char *outp = out;
+    bool isEmpty = true;
+
+    for(i=0;i<size;i++) {
+      switch(in[i]) {
+      case ' ':
+      case '\t':
+      case '\r':
+      case '\n':  // remove starting empty chars
+      case '\0':
+        if(!isEmpty)
+         *outp++ = in[i];
+        break;
+      default:
+        *outp++ = in[i];
+        isEmpty = false;
+        break;
+      } 
+    }
+    if(size) {
+      if(strncmp(in, "</body>", 7)==0) {
+        *outp++ = ')';
+        *outp++ = '\"';
+        *outp++ = ';';
+        *outp++ = '\r';
+        *outp++ = '\n';
+        *outp++ = 0;
+      }
+
+      fprintf(op, "%s", out);
+    }
+  }
+  fclose(fp);
+  fclose(op);
+}
