+++ OpenGarage/www_assets/publish_assets.sh
@@ -1,0 +1,6 @@
+./publish_assets views/ap_home.html html_ap_home
+./publish_assets views/sta_home.html html_sta_home
+./publish_assets views/sta_options.html html_sta_options
+./publish_assets views/sta_update.html html_sta_update
+./publish_assets views/sta_logs.html html_sta_logs
+./publish_assets views/portal.html portal
